Refactor: Introduce FileActionService for file operations and update DemoRunner to utilize it
